Add concept of Application. Reorganize build scripts to encapsulate modules dependencies

+++ scripts/generate_project_osx.sh
@@ -1,5 +1,9 @@
 cd ..
-rm -rf _build
+
+if [[ $# -eq 0 ]] ; then
+	rm -rf _build
+fi
+
 mkdir _build
 cd _build
 cmake .. -GXcode
